+++ tests/unit/odt/api/test_prsls.py
@@ -12,6 +12,7 @@ from ska_aaa_authhelpers.test_helpers import TEST_USER
 from ska_db_oda.persistence.domain.errors import ODANotFound
 
 from ska_oso_services.odt.api.prsls import ProposalProjectDetails
+from ska_oso_pdm.proposal.proposal import ProposalStatus
 from tests.unit.conftest import ODT_BASE_API_URL
 from tests.unit.util import TestDataFactory, assert_json_is_equal
 
@@ -27,7 +28,7 @@ class TestProjectCreationFromProposal:
     ):
         """ """
         uow_mock = mock.MagicMock()
-        proposal = TestDataFactory.proposal()
+        proposal = TestDataFactory.complete_proposal(status=ProposalStatus.SUBMITTED)
         uow_mock.prsls.get.return_value = proposal
         mock_uow().__enter__.return_value = uow_mock
 
@@ -48,7 +49,7 @@ class TestProjectCreationFromProposal:
     def test_project_status_is_created(self, mock_uow, mock_generate_project, client):
         """ """
         uow_mock = mock.MagicMock()
-        proposal = TestDataFactory.proposal()
+        proposal = TestDataFactory.complete_proposal(status=ProposalStatus.SUBMITTED)
         uow_mock.prsls.get.return_value = proposal
         mock_uow().__enter__.return_value = uow_mock
 
@@ -88,7 +89,7 @@ class TestProjectCreationFromProposal:
     def test_oda_error(self, mock_uow, client):
         """ """
         uow_mock = mock.MagicMock()
-        proposal = TestDataFactory.proposal()
+        proposal = TestDataFactory.complete_proposal(status=ProposalStatus.SUBMITTED)
         uow_mock.prsls.get.side_effect = IOError("test error")
         mock_uow().__enter__.return_value = uow_mock
         with pytest.raises(IOError):
@@ -108,7 +109,7 @@ class TestProposalAndProjectView:
     @mock.patch("ska_oso_services.odt.api.prsls.oda.uow")
     def test_proposal_without_project_returned_successfully(self, mock_uow, client):
         uow_mock = mock.MagicMock()
-        proposal = TestDataFactory.complete_proposal()
+        proposal = TestDataFactory.complete_proposal(status=ProposalStatus.SUBMITTED)
         uow_mock.prsls.query.return_value = [proposal]
 
         uow_mock.prjs.query.return_value = []
@@ -136,7 +137,7 @@ class TestProposalAndProjectView:
     @mock.patch("ska_oso_services.odt.api.prsls.oda.uow")
     def test_project_and_proposal_returned_successfully(self, mock_uow, client):
         uow_mock = mock.MagicMock()
-        proposal = TestDataFactory.complete_proposal()
+        proposal = TestDataFactory.complete_proposal(status=ProposalStatus.SUBMITTED)
         uow_mock.prsls.query.return_value = [proposal]
 
         project = TestDataFactory.project()
@@ -203,7 +204,7 @@ class TestProposalAndProjectView:
     @mock.patch("ska_oso_services.odt.api.prsls.oda.uow")
     def test_only_highest_version_returned_successfully(self, mock_uow, client):
         uow_mock = mock.MagicMock()
-        proposal_v1 = TestDataFactory.complete_proposal()
+        proposal_v1 = TestDataFactory.complete_proposal(status=ProposalStatus.SUBMITTED)
         proposal_v2 = copy.deepcopy(proposal_v1)
         proposal_v2.metadata.version = 2
 
@@ -243,5 +244,41 @@ class TestProposalAndProjectView:
                 prsl_id=proposal_v2.prsl_id,
                 prsl_version=2,
                 title=proposal_v2.info.title,
+            ).model_dump_json()
+        )
+
+    @mock.patch("ska_oso_services.odt.api.prsls.oda.uow")
+    def test_does_not_return_draft_proposals(self, mock_uow, client):
+        uow_mock = mock.MagicMock()
+        proposal_with_prj = TestDataFactory.complete_proposal(status=ProposalStatus.SUBMITTED)
+        proposal_with_prj.prsl_id = "prp-test-id"
+        proposal_draft = TestDataFactory.proposal()
+        uow_mock.prsls.query.return_value = [proposal_with_prj, proposal_draft]
+
+        project = TestDataFactory.project()
+        project.prsl_ref = proposal_with_prj.prsl_id
+        uow_mock.prjs.query.return_value = [project]
+
+        mock_uow().__enter__.return_value = uow_mock
+
+        resp = client.get(
+            f"{PRSLS_API_URL}/project-view",
+        )
+
+        assert resp.status_code == HTTPStatus.OK
+        result = resp.json()
+
+        assert len(result) == 1
+        # Dump to json then load so the datetimes are serialised
+        assert result[0] == json.loads(
+            ProposalProjectDetails(
+                prj_id=project.prj_id,
+                prj_version=1,
+                prsl_id=proposal_with_prj.prsl_id,
+                title=project.name,
+                prj_last_modified_by=project.metadata.last_modified_by,
+                prj_last_modified_on=project.metadata.last_modified_on,
+                prj_created_by=project.metadata.created_by,
+                prj_created_on=project.metadata.created_on,
             ).model_dump_json()
         )